feat(authentication-service): enhance logout functionality #main
Updated the logout functionality in the authentication service to support redirection
after logout. Added a new 'returnTo' query parameter to specify the URL for redirection.
Also, added 'postLogoutRedirectUris' property to the auth client model. This is an
array of URLs to which a user can be redirected after logging out.

The 'logoutRedirectUrl' property is now deprecated and 'postLogoutRedirectUris' should
be used instead.

Added a new utility function 'isAllowedUri' to validate whether a given URI matches
any pattern in a list of URIs. This is used to ensure that the 'returnTo' URL matches
one of the 'postLogoutRedirectUris' specified for the auth client.

Updated the logout controller to perform the necessary validations and handle redirection
as per the new changes.

Also added unit tests for the new utility function 'isAllowedUri'.